requirements.txt
scipy==1.11.3
PyYAML==6.0.1
datasets==2.14.6
transformers==4.35.0
accelerate==0.24.1
trl==0.7.2
bitsandbytes==0.41.1
peft==0.6.0
einops==0.7.0
wandb==0.15.12